requirements.txt
anthropic>=0.40.0
python-dotenv>=1.0.0
requests>=2.32.0
beautifulsoup4>=4.12.0
feedparser>=6.0.0
fpdf2>=2.8.0
schedule>=1.2.0
colorama>=0.4.6
pytest>=8.0.0
